Refactor: rename whole polyline_trajectory_follower directory + rename polyline_trajectory_follower file to follow_trajectory
Change include paths and include guards
