+++ openscenario/openscenario_interpreter/include/openscenario_interpreter/syntax/follow_trajectory_action.hpp
@@ -22,7 +22,7 @@
 #include <openscenario_interpreter/syntax/trajectory_following_mode.hpp>
 #include <openscenario_interpreter/syntax/trajectory_ref.hpp>
 #include <pugixml.hpp>
-#include <traffic_simulator/behavior/polyline_trajectory_follower/polyline_trajectory_follower.hpp>
+#include <traffic_simulator/behavior/follow_trajectory/follow_trajectory.hpp>
 
 namespace openscenario_interpreter
 {
